requirements.txt
Django==5.2.5
djangorestframework==3.16.0
django-cors-headers==4.7.0
django-extensions==3.2.3
django-crispy-forms==2.4
crispy-bootstrap5==2025.6
django-imagekit==5.0.0
django-cleanup==9.0.0
python-decouple==3.8
django-filter==25.1
psycopg2-binary==2.9.10
redis==5.2.0
cloudinary==1.44.1
celery==5.5.3
whitenoise==6.9.0
gunicorn==23.0.0
sentry-sdk==2.35.0

